alif_testpl/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.11
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.1
# azure-cognitiveservices-speech>=1.34.0
openai>=1.10.0
aiofiles>=23.2.1
slowapi>=0.1.9
email-validator>=2.1.0
requests>=2.31.0
python-docx>=1.1.0
pypdf>=4.0.0
chardet>=5.2.0


# TestAI Dependencies
pdfplumber>=0.10.3
# pytesseract>=0.3.10
pillow>=10.1.0
# python-magic>=0.4.27
langdetect>=1.0.9
